Remove cash and replace with shelljs-core
Cash is out of date and hard to patch for injecting our own filesystem
into. This change removes cash and introduces our patched shelljs-core
module to implement some common methods in shelljs.

Refactored the console to make adding methods a little easier and
reduce the filesize of the main Console application.